+++ GCR Project/GCR Project.vcxproj.filters
@@ -21,7 +21,7 @@
     <ClInclude Include="File.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
-    <ClInclude Include="EmployeeMananger.h">
+    <ClInclude Include="EmployeeManager.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
     <ClInclude Include="Employee.h">
